Pending tasks
git-svn-id: https://svn.code.sf.net/p/kmetronome/code/trunk@114 ea54cc14-8072-40c6-bac7-338dc0ed4500

+++ src/drumgrid.cpp
@@ -232,8 +232,15 @@ void DrumGrid::writePattern(const QString& name)
 
 void DrumGrid::removePattern(const QString& name)
 {
-    KGlobal::config()->deleteGroup(QSTR_PATTERN+name);
-    KGlobal::config()->sync();
+    if ( KMessageBox::questionYesNo (this,
+            i18n("Do you want to remove the current pattern?"),
+            i18n("Remove Pattern"),
+            KStandardGuiItem::yes(),
+            KStandardGuiItem::no(),
+            "removepattern") == KMessageBox::Yes) {
+        KGlobal::config()->deleteGroup(QSTR_PATTERN+name);
+        KGlobal::config()->sync();
+    }
 }
 
 QStringList DrumGrid::patterns()
@@ -254,7 +261,9 @@ void DrumGrid::patternChanged(int /*idx*/)
 
 void DrumGrid::savePattern()
 {
-    writePattern(m_ui->patternCombo->currentText());
+    QString newName = m_ui->patternCombo->currentText();
+    writePattern(newName);
+    m_ui->patternCombo->addItem(newName);
 }
 
 void DrumGrid::removePattern()
